Safari-609.1.11.1/LayoutTests/http/tests/xmlhttprequest/access-control-preflight-async-not-supported-expected.txt

CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-preflight-denied-xsrf.php?test=access-control-preflight-async-non-supported due to access control checks.
PASS: Request successfully blocked.
